ConvertToUI updated to include session type in String. MakePDF and ConvertToUI updated to reuse the same code
